+++ src/managers/SettingsManager.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+class SettingsManager {
+public:
+    static bool ShowDailyButton;
+    static bool ShowWeeklyButton;
+    static bool ShowSendOnlyButton;
+};