+++ changelog.md
@@ -1,3 +1,5 @@
 # v1.0.0
-- Improved performance (very slightly) & clarified the read me & some popups' text
+- Added toggleable buttons to set the level for daily/weekly & to only send the level
+- Improved performance (very slightly) & clarified the read me and some popups' text
+- Added an ID to the difficulty face sprite
 - Mod now disables itself when not on a GDPS